take 3
bench: 2527776

+++ src/transposition_table.cpp
@@ -84,10 +84,10 @@ void TranspositionTable::store(uint64_t zobrist, int value, int static_eval, int
         (depth != DEPTH_QSEARCH && flag == TFlag::EXACT))
     {
         // add move if the old entry didn't hold the same position or if the new move is better
-        if (entry->zobrist_hash != zobrist || move != Move::NO_MOVE)
+        if (entry->zobrist_hash != uint32_t(zobrist) || move != Move::NO_MOVE)
             entry->move = move.move();
 
-        entry->zobrist_hash = zobrist;
+        entry->zobrist_hash = uint32_t(zobrist);
         entry->value = value;
         entry->static_eval = static_eval;
         entry->depth_tflag = (static_cast<uint8_t>(depth) << 2) | (static_cast<uint8_t>(flag));
